Apply calibration filters only to calibration
dataFrame was passed to calibration calculation, which is then filtered
and date exclusions applied. These then carry through to the PP period
(as only a pointer to the dataframe object is passed, rather than the
data itself). To fix this dataFrame is copied at start of calibration
period and reset at the end.